% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_lags}
\alias{add_lags}
\title{Add leads and lags to a datafranme, with panel ID variables
id and date}
\usage{
add_lags(df, vars, max.p = 3, lags = 4, add_leads = FALSE, sort_df = TRUE)
}
\arguments{
\item{sort_df}{}
}
\description{
Add leads and lags to a datafranme, with panel ID variables
id and date
}
