% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkCase.r
\name{all.lower}
\alias{all.lower}
\title{all.lower}
\usage{
\method{all}{lower}(string)
}
\arguments{
\item{string}{Character vector of strings to check cases}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all lower case
}
\details{
Checks if strings are all lower case. This is a wrapper for \code{find.case('text', 'lower')}.
}
\examples{
toCheck <- c('BIG', 'little', 'Mixed', 'BIG WITH SPACE', 'little with space', 'MIXED with SPACE')
all.lower(toCheck)
}
\author{
Jared P. Lander
}
\seealso{
find.case all.upper
}

