% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkCase.r
\name{all.upper}
\alias{all.upper}
\title{all.upper}
\usage{
\method{all}{upper}(string)
}
\arguments{
\item{string}{Character vector of strings to check cases}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all upper case
}
\details{
Checks if strings are all upper case.  This is a wrapper for \code{find.case('text', 'upper')}.
}
\examples{
all.upper(toCheck)
}
\author{
Jared P. Lander
}
\seealso{
find.case all.lower
}

