% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlot.r
\name{autoplot.acf}
\alias{autoplot.acf}
\title{autoplot.acf}
\usage{
\method{autoplot}{acf}(object, xlab = x, ylab = y, title = sprintf("\%s Plot", y), ...)
}
\arguments{
\item{object}{An \code{\link{acf}} object.}

\item{xlab}{X-axis label.}

\item{ylab}{y-axis label.}

\item{title}{Graph title.}

\item{\dots}{Further arguments.}
}
\value{
A ggplot object.
}
\description{
Plot acf objects
}
\details{
Plot acf (and pacf) objects.
}
\examples{

autoplot(acf(sunspot.year, plot=FALSE))
autoplot(pacf(sunspot.year, plot=FALSE))

}
\author{
Jared P. Lander
}
