% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.flip.r
\name{binary.flip}
\alias{binary.flip}
\title{binary.flip}
\usage{
binary.flip(x)
}
\arguments{
\item{x}{A vector of 0/1 numbers.}
}
\value{
X with 0's flipped to 1's and 1's flipped to 0's
}
\description{
Flip binary numbers
}
\examples{

binary.flip(c(1,1,0,1,0,0,1))

}
\author{
Jared P. Lander
}
