% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFormula.r
\name{build.formula}
\alias{build.formula}
\title{Formula Builder}
\usage{
build.formula(lhs, rhs)
}
\arguments{
\item{lhs}{Character vector for left side of formula}

\item{rhs}{Character vector for right side of formula}
}
\value{
A formula object
}
\description{
Formula Builder
}
\details{
Builds a formula easily given the left and right hand sides.  Right now it only handles additive formulas and not interactions unless that is specified in the character.
}
\examples{

build.formula("Y", "X")
build.formula(c("Y", "Z"), "X")
build.formula("Z", c("X", "Q"))
build.formula(c("Y", "Z"), c("X", "Q"))

}
\seealso{
formula as.formula
}
\author{
Jared P. Lander www.jaredlander.com
}
