% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classCheck.r
\name{classDdf}
\alias{classDdf}
\alias{classdf}
\title{classdf}
\usage{
classDdf(data, cols)
}
\arguments{
\item{data}{\code{link{data.frame}} that is to be inspected.}

\item{cols}{The columns (named or numeric) to be included in the check.}
}
\value{
A vector detailing the class of each column.
}
\description{
Get class information for each column in a \code{\link{data.frame}}.
}
\details{
Get class information for each column in a \code{\link{data.frame}}.
}
\examples{
classDF(CO2)
classDF(iris)
classDF(mtcars)
}
\author{
Jared P. Lander
}

