% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.r
\name{constant}
\alias{constant}
\title{constant}
\usage{
constant(n = 1)
}
\arguments{
\item{n}{The value to return}
}
\value{
A function that when used simply returns n.
}
\description{
Helper function for imputing constants
}
\details{
Returns a function that always returns the value of n.
}
\examples{
constant(4)(1:10)

theDF <- data.frame(A=1:10, B=1:10, C=1:10)
theDF[c(1, 4, 6), c(1)] <- NA
theDF[c(3, 4, 8), c(3)] <- NA
simple.impute(theDF, constant(4))

}
\author{
Jared P. Lander
}
