% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{do.call.rbind}
\alias{do.call.cbind}
\alias{do.call.rbind}
\title{Quicker rbind / cbind}
\usage{
do.call.rbind(lst)
}
\arguments{
\item{lst}{List of xts objects}
}
\value{
combined single xts object
}
\description{
Combines list of xts objects into one large xts object.
from Dominik and GSee at
http://stackoverflow.com/questions/12028671/merging-a-large-list-of-xts-objects
}
\examples{
l <- lapply(Sys.Date()-6000:1, function(x) {N=60*8;list(rnorm(N),as.POSIXct(x)-seq(N*60,1,-60))})
system.time(l1 <- do.call.rbind(l))
}

