% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecobrewer.pal.src.R
\name{ecobrewer.pal}
\alias{ecobrewer.pal}
\title{Recycle brewer.pal() palette colors}
\usage{
ecobrewer.pal(n, name)
}
\arguments{
\item{n}{number of colors to pass to brewer.pal()}

\item{name}{name of color palette to extract colors from}
}
\description{
Wrapper around brewer.pal from RColorBrewer to reuse colors
}
\examples{
ecobrewer.pal(34, "Set1")
}
