% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCase.r
\name{find.case}
\alias{find.case}
\title{find.case}
\usage{
find.case(string, case = c("upper", "lower", "mixed", "numeric"))
}
\arguments{
\item{string}{Character vector of strings to check cases}

\item{case}{Whether checking for upper or lower case}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all upper or all lower case
}
\details{
Checks if strings are all upper or all lower case.  If string is all numbers it returns TRUE.
}
\examples{
toCheck <- c('BIG', 'little', 'Mixed', 'BIG WITH SPACE', 'little with space', 'MIXED with SPACE')
find.case(toCheck, 'upper')
find.case(toCheck, 'lower')
}
\seealso{
upper.case lower.case numeric.case mixed.case
}
\author{
Jared P. Lander
}
