% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{fix_mat}
\alias{fix_mat}
\title{Fix a Matrix to Ensure Positive Semi-Definiteness}
\usage{
fix_mat(X, ridge = 1e-20)
}
\arguments{
\item{X}{A symmetric matrix to be fixed.}

\item{ridge}{A small positive value used as the minimum eigenvalue. Default is \code{1e-20}.}
}
\value{
A symmetric positive semi-definite matrix.
}
\description{
This function adjusts a given matrix to ensure it is positive semi-definite (PSD).
If the matrix has negative eigenvalues, they are replaced with a small positive value
(ridge adjustment) to make the matrix PSD.
}
\details{
The function checks the eigenvalues of the input matrix. If all eigenvalues are positive,
the matrix is returned as is. If there are negative eigenvalues, the function adjusts them
to ensure the matrix is PSD. If the smallest positive eigenvalue is smaller than the ridge,
a ridge adjustment is applied.

If the matrix cannot be fixed (i.e., it still has negative eigenvalues after adjustment),
the function throws an error.
}
\examples{
# Example usage:
mat <- matrix(c(2, -1, -1, 2), nrow = 2)
fixed_mat <- fix_mat(mat)

}
