% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlot.r
\name{fortify.ts}
\alias{fortify.ts}
\title{fortify.ts}
\usage{
\method{fortify}{ts}(model, data = NULL, name = as.character(m[[2]]), ...)
}
\arguments{
\item{model}{A \code{\link{ts}} object.}

\item{data}{A vector of the same length of \code{x} that specifies the time component of each element of \code{x}.}

\item{name}{Character specifying the name of x if it is to be different that the variable being inputed.}

\item{\dots}{Further arguments.}
}
\value{
\code{\link{data.frame}} for plotting with ggplot.
}
\description{
Fortify a ts object.
}
\details{
Prepares a ts object for plotting with ggplot.
}
\examples{

fortify(sunspot.year)

}
\author{
Jared P. Lander
}
