% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dateTime.src.R
\name{get_dateTime}
\alias{get_dateTime}
\title{Sys.time() wrapper parse}
\usage{
get_dateTime()
}
\description{
Parses Sys.time() for a filename-friendly timestamp
}
\examples{
get_dateTime() 
[1] "2019-01-01_000001"
}
