% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.r
\name{left}
\alias{left}
\title{Grabs the left side of a data set}
\usage{
left(x, c = 5L, ...)
}
\arguments{
\item{x}{The data}

\item{c}{Number of columns to show}

\item{\dots}{Arguments passed on to other functions}
}
\value{
... The left side of the data set that was requested.  The size depends on c.
}
\description{
Display the left side of a rectangular data set
}
\details{
Displays the left side of a rectangular data set.

This is a wrapper function for \code{\link{corner}}
}
\examples{

data(diamonds)
head(diamonds)      # displays all columns
left(diamonds)    # displays all rows and only the first 5 columns

}
\seealso{
\code{\link{head}} \code{\link{tail}} \code{\link{corner}} \code{\link{topright}} \code{\link{bottomleft}} \code{\link{bottomright}} \code{\link{topleft}} \code{\link{right}}
}
\author{
Jared P. Lander www.jaredlander.com
}
\keyword{corner}
\keyword{display}
\keyword{head}
\keyword{subsection}
\keyword{tail}
\keyword{view}
