% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.src.R
\name{load_library_list}
\alias{load_library_list}
\title{load a list of libraries}
\usage{
load_library_list(library_list)
}
\arguments{
\item{library_list}{character list of packages to load}
}
\description{
take a list of installed packages and load them all
}
\examples{
load_library_list(c("data.table", "Biostrings"))
load_library_list(list_of_packages)
}
