% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.r
\name{load_packages}
\alias{load_packages}
\title{load_packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{A `character` vector of packages to be installed}
}
\value{
Nothing, loads packages
}
\description{
Loads multiple packages
}
\details{
Allows the user to load multiple packages with one line of code. Delivers a message saying which packages have been loaded. If a user requests packages that are not installed there is an error.
}
\examples{

load_packages('ggplot2')
load_packages(c('ggplot2', 'dplyr'))
}
\author{
Jared P. Lander
}
