% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCase.r
\name{lower.case}
\alias{lower.case}
\title{lower.case}
\usage{
lower.case(string)
}
\arguments{
\item{string}{Character vector of strings to check cases}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all lower case
}
\details{
Checks if strings are all lower case. This is a wrapper for \code{find.case('text', 'lower')}.  If string is all numbers it returns TRUE.
}
\examples{
toCheck <- c('BIG', 'little', 'Mixed', 'BIG WITH SPACE', 'little with space', 'MIXED with SPACE')
lower.case(toCheck)
}
\seealso{
find.case upper.case mixed.case numeric.case
}
\author{
Jared P. Lander
}
