% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCase.r
\name{numeric.case}
\alias{numeric.case}
\title{numeric.case}
\usage{
numeric.case(string)
}
\arguments{
\item{string}{Character vector of strings to check cases}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all numbers or spaces
}
\details{
Checks if strings are all numbers and spaces. This is a wrapper for \code{find.case('text', 'numeric')}.
}
\examples{
toCheck <- c('BIG', 'little', 'Mixed', 'BIG WITH SPACE', 
     'little with space', 'MIXED with SPACE', '17')
numeric.case(toCheck)
}
\seealso{
find.case upper.case lower.case numeric.case
}
\author{
Jared P. Lander
}
