% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_poly_sum_lags.R
\name{plotdf.het}
\alias{plotdf.het}
\title{Get a prediction df for a lagged response function with heterogeneity}
\usage{
plotdf.het(
  df.reg,
  m,
  treat.var = "precip_p",
  min = 0,
  max = 300,
  ref = 0,
  het.var = "het",
  lags = 4
)
}
\arguments{
\item{lags}{}
}
\description{
Get a prediction df for a lagged response function with heterogeneity
}
