% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.src.R
\name{random_DNA}
\alias{random_DNA}
\title{random DNA sequence generator}
\usage{
random_DNA(length, is_RNA = FALSE, seed = NULL)
}
\arguments{
\item{length}{length of DNA sequence to be generated}

\item{is_RNA}{defaults to FALSE}

\item{seed}{random seed; defaults to NULL}
}
\description{
generate a random string of DNA (or RNA) bases
}
\examples{
random_DNA(12, is_RNA = TRUE, seed = 123)
}
