% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.src.R
\name{read_fasta}
\alias{read_fasta}
\title{fasta batch Biostrings}
\usage{
read_fasta(files, is_RNA = FALSE, output_list = FALSE)
}
\arguments{
\item{files}{path or list of paths to fasta files}

\item{is_RNA}{logical argument to use RNAStringSet - default is FALSE}

\item{output_list}{logical argument to concatenate list of XStringSets - default is FALSE}
}
\description{
extract DNAStringSet data from a batch of fasta files: requires package Biostrings
}
\examples{
read_fasta(files = c("path/to/example1.fa", "path/to/example2.fa"), is_RNA = FALSE, output_list = FALSE)
}
