% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genbank.src.R
\name{read_genbank}
\alias{read_genbank}
\title{simple genbank file reader}
\usage{
read_genbank(files)
}
\arguments{
\item{files}{path or list of paths to genbank files}
}
\description{
extract DNAStringSet data from genbank files
}
\examples{
random_DNA(c("example1.gb", "example2.gb"))
}
