% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.r
\name{reclass}
\alias{reclass}
\alias{reclass<-}
\title{reclass}
\usage{
reclass(x, value)

reclass(x) <- value
}
\arguments{
\item{x}{The x getting the new class}

\item{value}{The new class}
}
\value{
The original x with the class containing \code{value} in addition to the previous class(es)
}
\description{
Adds a class to an x.
}
\details{
Adds a class to an x by putting the new class at the front of the vector of classes for the x.
}
\examples{
theDF <- data.frame(A=1:10, B=1:10)
reclass(theDF) <- 'newclass'
class(theDF)
theDF <- reclass(theDF, 'another')
class(theDF)

}
\author{
Jared P. Lander
}
