% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_zero_NA.src.R
\name{tab_zero_NA}
\alias{tab_zero_NA}
\title{data.table replace NAs}
\usage{
tab_zero_NA(data, replacement = 0, return_data.table = TRUE)
}
\arguments{
\item{data}{either data.table or data.frame object}

\item{replacement}{value to replace NAs with}

\item{return_data.table}{logical argument for table type (defaults to TRUE)}
}
\description{
replaces all NAs in a data.table with a set value
}
\examples{
tab_zero_NA(table, replacement = "0", return_data.frame = TRUE)
}
