% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.single.R
\name{timeSingle}
\alias{timeSingle}
\title{timeSingle}
\usage{
timeSingle(
  string = "Time difference",
  startTime,
  endTime = Sys.time(),
  sep = ":"
)
}
\arguments{
\item{string}{string of what was timed}

\item{startTime}{"POSIXct" "POSIXt" object, usually from \code{\link{Sys.time}}}

\item{endTime}{"POSIXct" "POSIXt" object, usually from \code{\link{Sys.time}}}

\item{sep}{string, usually character that is used as the separator between user prompt and time difference}
}
\value{
prompt_string string user prompt with time difference
}
\description{
Convenience function that takes in a time object and calculates a difference
with a user specified prompt
}
\examples{

x <- 3.14
strt <- Sys.time()
sq <- x ** 2
timeSingle('Squaring value', strt)

}
\author{
Daniel Y. Chen
}
