% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-package.r
\docType{package}
\name{useful-package}
\alias{useful}
\alias{useful-package}
\title{Helper functions}
\description{
A collection of handy, helper functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaredlander/useful}
  \item Report bugs at \url{https://github.com/jaredlander/useful/issues}
}

}
\author{
\strong{Maintainer}: Jared P. Lander \email{packages@jaredlander.com}

Other contributors:
\itemize{
  \item Nicholas Galasinao \email{nicholas.galasinao@gmail.com} [contributor]
  \item Joshua Kraut \email{joshua.kraut@gmail.com} [contributor]
  \item Daniel Chen \email{dcheny@gmail.com} [contributor]
}

}
\keyword{internal}
