% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{vcov_psd}
\alias{vcov_psd}
\title{Ensure Positive Semi-Definiteness of Variance-Covariance Matrix}
\usage{
vcov_psd(object, ...)
}
\arguments{
\item{object}{A model object for which the variance-covariance matrix is to be fixed.}

\item{...}{Additional arguments passed to \code{vcov}.}
}
\value{
A symmetric positive semi-definite variance-covariance matrix.
}
\description{
This function adjusts the variance-covariance matrix of a model object to ensure it is
positive semi-definite (PSD).
}
\details{
This function retrieves the variance-covariance matrix of the given model object using
\code{vcov}, and then applies \code{fix_mat} to ensure the matrix is PSD.
}
\examples{
# Example usage:
# Assuming `model` is a fitted model object:
# fixed_vcov <- vcov_psd(model)

}
