% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlot.r
\name{vplayout}
\alias{vplayout}
\title{vplayout}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{The x cell of the viewport to push into.}

\item{y}{The y cell of the viewport to push into.}
}
\value{
An R object of class viewport.
}
\description{
Viewport
}
\details{
Creates viewport for pushing ggplot objects to parts of a console.
}
\examples{

library(ggplot2)
library(grid)

}
\author{
Jared P. Lander
}
