% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMatrix.r
\name{build.y}
\alias{build.y}
\title{build.y}
\usage{
build.y(formula, data)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data.frame}
}
\value{
The y object from a formula and data
}
\description{
Build the y object from a formula and data
}
\details{
Given a formula and a data.frame build the y object
}
\examples{
require(ggplot2)
head(mpg)
head(build.y(hwy ~ class + cyl + year, data=mpg))

}
\author{
Jared P. Lander
}
