% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{build_formula_poly}
\alias{build_formula_poly}
\title{Build a formula for fixest}
\usage{
build_formula_poly(
  yvar,
  treat,
  control,
  poly_treat,
  poly_control,
  leads,
  lags,
  FE,
  interact = NULL,
  manual_treat = NULL
)
}
\arguments{
\item{manual_treat}{}
}
\description{
Build a formula for fixest
}
