% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{gen_ff}
\alias{gen_ff}
\title{Generate a string for a given variable}
\usage{
gen_ff(leads, lags, treatvar, interact = NULL)
}
\arguments{
\item{interact}{}
}
\description{
Generate a string for a given variable
}
