% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlot.r
\name{plot.acf}
\alias{plot.acf}
\title{plot.acf}
\usage{
\method{plot}{acf}(x, ...)
}
\arguments{
\item{x}{An ACF object}

\item{\dots}{Arguments passed on to autoplot}
}
\value{
A ggplot2 object
}
\description{
Plotting an ACF object
}
\details{
This function has been deprecated in favor of autoplot
}
\author{
Jared P. Lander
}
