% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsPlot.r
\name{plot}
\alias{plot}
\title{plot}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{\dots}{Further arguments}
}
\value{
A plot
}
\description{
Overwitten plot generic so that plot.acf can be defined in this package
}
\details{
Overwitten plot generic so that plot.acf can be defined in this package
}
\author{
Jared P. Lander
}
