% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_poly.R
\name{plot_rf_poly}
\alias{plot_rf_poly}
\title{Plot the output of a predict_poly call}
\usage{
plot_rf_poly(tab, xvar = "temp")
}
\arguments{
\item{tab}{: the output of predict_poly()}
}
\value{
a ggplot object with some nice color bars
}
\description{
Plot the output of a predict_poly call
}
