% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lags_separately.R
\name{predict_lags_separately}
\alias{predict_lags_separately}
\title{Title: Extract lagged resposne function values separately for each lag}
\usage{
predict_lags_separately(
  m,
  var,
  min,
  max,
  ref,
  leads,
  lags,
  ci_level = 95,
  step.length = 1,
  id.col = NULL,
  xvar_name = "temp"
)
}
\arguments{
\item{leads}{number of leads in the model}

\item{lags}{number of lags in the model}

\item{xvar_name}{}
}
\description{
Title: Extract lagged resposne function values separately for each lag
}
