% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genbank.src.R
\name{read_genbank}
\alias{read_genbank}
\title{genbank batch Biostrings}
\usage{
read_genbank(files, output_list = FALSE)
}
\arguments{
\item{files}{path or list of paths to genbank files}

\item{output_list}{logical argument to concatenate list of XStringSets - default is FALSE}
}
\description{
extract DNAStringSet data from batch of genbank files: requires package Biostrings
}
\examples{
random_DNA(files = c("path/to/example1.gb", "path/to/example2.gb"), output_list = FALSE)
}
