% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{test_het_coefs}
\alias{test_het_coefs}
\title{Test for equality of coefficients across some margin of heterogeneity}
\usage{
test_het_coefs(m, poly, lags)
}
\arguments{
\item{m}{: fixest object with interacted treatment variables}

\item{poly}{: the order of the polynomial in the treatment}

\item{lags}{the number of lags of the treatment variabe}
}
\description{
Test for equality of coefficients across some margin of heterogeneity
}
