% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCase.r
\name{upper.case}
\alias{upper.case}
\title{upper.case}
\usage{
upper.case(string)
}
\arguments{
\item{string}{Character vector of strings to check cases}
}
\value{
A vector of TRUE AND FALSE
}
\description{
Checks if strings are all upper case
}
\details{
Checks if strings are all upper case.  This is a wrapper for \code{find.case('text', 'upper')}.  If string is all numbers it returns TRUE.
}
\examples{
toCheck <- c('BIG', 'little', 'Mixed', 'BIG WITH SPACE', 'little with space', 'MIXED with SPACE')
upper.case(toCheck)
}
\seealso{
find.case lower.case mixed.case numeric.case
}
\author{
Jared P. Lander
}
