% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-package.R
\docType{package}
\name{useful-package}
\alias{useful}
\alias{useful-package}
\title{useful: Some functions I'm always using in empirical climate econometrics research}
\description{
A bunch of helpful functions for running fixed effects regressions, plotting their outputs. Heavily dependent on the feols package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TomBearpark/useful}
  \item Report bugs at \url{https://github.com/TomBearpark/useful/issues}
}

}
\author{
\strong{Maintainer}: Thomas Bearpark \email{t.bearpark@gmail.com}

}
\keyword{internal}
